Add support for config file
