+++ src/config.example.json
@@ -1,0 +1,8 @@
+{
+    "instrument_name": "SXD",
+    "wavelength_range": [0.5, 10.0],
+    "extents": [-17, 17, -7, 17, 0, 33],
+    "nbins": 300,
+    "temperature": 50.0,
+    "alpha": 0.3e-3
+}
